Add limited concurrency back.
Even though we are limiting the number of threads actively downloading
data from nix at any one time, if many threads are trying to write to
the database, they will all load the nix data from disk into memory
without sharing and be blocked in the queue. This was causing crazy
memory usage. Using limitedConcurrency is a simple fix